+++ frontend/src/routes/_dashboard.tsx
@@ -3,7 +3,7 @@ import {
   Link,
   Outlet,
   createFileRoute,
-  useNavigate
+  useNavigate,
 } from "@tanstack/react-router";
 import {
   AlarmCheck,
@@ -21,7 +21,7 @@ import {
   Search,
   Send,
   Settings,
-  Twitter
+  Twitter,
 } from "lucide-react";
 import { apiClient } from "~/api/client";
 import { useAuthUser } from "~/components/helper/use-auth-user";
@@ -30,15 +30,16 @@ import { Input } from "~/components/ui/input";
 import {
   Menubar,
   MenubarContent,
+  MenubarContentItem,
   MenubarItem,
   MenubarMenu,
-  MenubarTrigger
+  MenubarTrigger,
 } from "~/components/ui/menubar";
 import {
   Sheet,
   SheetContent,
   SheetHeader,
-  SheetTrigger
+  SheetTrigger,
 } from "~/components/ui/sheet";
 import { userKeys } from "~/key-factories";
 import { cn } from "~/lib/utils";
@@ -53,7 +54,7 @@ export const Route = createFileRoute("/_dashboard")({
       </main>
       <Footer />
     </div>
-  )
+  ),
 });
 
 function Header() {
@@ -68,20 +69,20 @@ function Header() {
       const csrfToken = getCookie("csrftoken");
       const { error } = await apiClient.DELETE("/api/auth/logout/", {
         headers: {
-          "X-CSRFToken": csrfToken
-        }
+          "X-CSRFToken": csrfToken,
+        },
       });
       if (error) {
         return error;
       }
 
       queryClient.removeQueries({
-        queryKey: userKeys.authedUser
+        queryKey: userKeys.authedUser,
       });
       deleteCookie("csrftoken");
       navigate({ to: "/login" });
       return null;
-    }
+    },
   });
 
   if (!user) {
@@ -225,23 +226,23 @@ const socialLinks = [
   {
     name: "Feedback",
     url: " https://github.com/zane-ops/zane-ops/discussions",
-    icon: <Send width={socialLinksIconWidth} />
+    icon: <Send width={socialLinksIconWidth} />,
   },
   {
     name: "Docs",
     url: "https://github.com/zane-ops/zane-ops/blob/main/docs.md",
-    icon: <BookOpen width={socialLinksIconWidth} />
+    icon: <BookOpen width={socialLinksIconWidth} />,
   },
   {
     name: "Contribute",
     url: "https://github.com/zane-ops/zane-ops/blob/main/CONTRIBUTING.md",
-    icon: <HeartHandshake width={socialLinksIconWidth} />
+    icon: <HeartHandshake width={socialLinksIconWidth} />,
   },
   {
     name: "Twitter",
     url: "https://twitter.com/zaneopsdev",
-    icon: <Twitter width={socialLinksIconWidth} />
-  }
+    icon: <Twitter width={socialLinksIconWidth} />,
+  },
 ];
 
 function Footer() {
@@ -262,26 +263,5 @@ function Footer() {
         ))}
       </div>
     </>
-  );
-}
-
-type MenubarContentItemProps = {
-  icon: React.ElementType;
-  text: string;
-  className?: string;
-};
-
-export function MenubarContentItem({
-  icon: Icon,
-  text,
-  className
-}: MenubarContentItemProps) {
-  return (
-    <MenubarItem
-      className={cn("flex pr-4 w-full gap-2 cursor-pointer", className)}
-    >
-      {Icon && <Icon className={cn("w-4 opacity-50", className)} />}
-      {text}
-    </MenubarItem>
   );
 }
